hook Settings`... values to behavior changes
__main__.py
-----------

initializing termshell object `shell` will set the default Pygments style.
We should pick up that value to stor in Settings`PygmentsStyle

Right now, `from_python()` converts a boolean into 0 or 1. I don't know
if that is what it _should_ do, bit since that is the behavior now, we
need to retrieve it as an integer or fix up `is_true()` so that it
distinguishes 0 from 1. I went with Mathics' current behavior.

If we improve either `from_python()` or `is_true` we may revisit this
code.

termshell.py
------------

When changing styles we need to recreate a terminal formatter.
`change_pygments_style()` was added to do this.

settings.py
-----------

remove since we'll do this differently